Update PF_RING statistics Move linuxcommon_get_dev_stats to linux helpers, Including the old header file caused issues for PF_RING
