+++ lib/format_linux_xdp.c
@@ -539,7 +539,7 @@ static int linux_xdp_setup_xdp(libtrace_t *libtrace) {
     }
 
     // get the initial device stats
-    if (linuxcommon_get_dev_statistics(XDP_FORMAT_DATA->cfg.ifname, &XDP_FORMAT_DATA->cfg.stats) != 0) {
+    if (linux_get_dev_statistics(XDP_FORMAT_DATA->cfg.ifname, &XDP_FORMAT_DATA->cfg.stats) != 0) {
         XDP_FORMAT_DATA->cfg.stats.if_name[0] = 0;
     }
 
@@ -1393,7 +1393,7 @@ static void linux_xdp_get_stats(libtrace_t *libtrace, libtrace_stat_t *stats) {
 
     /* If we have the initial interface stats get the current and calculate the dropped packets */
     if (XDP_FORMAT_DATA->cfg.stats.if_name[0] != 0) {
-        if (linuxcommon_get_dev_statistics(XDP_FORMAT_DATA->cfg.ifname, &dev_stats) == 0) {
+        if (linux_get_dev_statistics(XDP_FORMAT_DATA->cfg.ifname, &dev_stats) == 0) {
             stats->dropped += (dev_stats.rx_drops - XDP_FORMAT_DATA->cfg.stats.rx_drops);
             stats->dropped_valid = 1;
             /* Received comes from the BPF program i.e. kernel
